Fix Armour (vests)
