+++ Addons/Armour/hpp/XtdGearInfos.hpp
@@ -1,0 +1,44 @@
+class XtdGearInfos {
+	class CfgWeapons {
+		class ONI_Recon_Armour_SF {
+			model = "ONI_SF_Armour";
+			camo = "A";
+		};
+		class ONI_Recon_Armour_SF_Medic {
+			model = "ONI_SF_Armour";
+			camo = "B";
+		};
+		class ONI_Recon_Armour_SF_Mx {
+			model = "ONI_SF_Armour";
+			camo = "C";
+		};
+		class ONI_Recon_Armour_SF_Sniper {
+			model = "ONI_SF_Armour";
+			camo = "D";
+		};
+		class ONI_Recon_Armour_SF_Scout {
+			model = "ONI_SF_Armour";
+			camo = "E";
+		};
+		class ONI_Recon_Armour_SF_Demo {
+			model = "ONI_SF_Armour";
+			camo = "F";
+		};
+		class ONI_Recon_Armour_SF_Carmichael {
+			model = "ONI_SF_Armour";
+			camo = "AG";
+		};
+		class ONI_Recon_Armour_SF_Hex {
+			model = "ONI_SF_Armour";
+			camo = "H";
+		};
+		class ONI_Recon_Armour_SF_Celt {
+			model = "ONI_SF_Armour";
+			camo = "sheepshagger";
+		};
+		class ONI_Recon_Armour_SF_Storm {
+			model = "ONI_SF_Armour";
+			camo = "purple";
+		};
+	};
+};
